+++ src/components/badge/badge.test.tsx
@@ -1,0 +1,7 @@
+import { Badge } from '.';
+
+describe('Badge', () => {
+  it('is truthy', () => {
+    expect(Badge).toBeTruthy();
+  });
+});