+++ src/components/buttons/buttons.test.tsx
@@ -1,7 +1,0 @@
-import { Buttons } from '.';
-
-describe('Buttons', () => {
-  it('is truthy', () => {
-    expect(Buttons).toBeTruthy();
-  });
-});