+++ src/components/card/card.test.tsx
@@ -1,0 +1,7 @@
+import { Card } from '.';
+
+describe('Card', () => {
+  it('is truthy', () => {
+    expect(Card).toBeTruthy();
+  });
+});