+++ src/components/divider/divider.test.tsx
@@ -1,0 +1,7 @@
+import { Divider } from '.';
+
+describe('Divider', () => {
+  it('is truthy', () => {
+    expect(Divider).toBeTruthy();
+  });
+});