+++ src/components/input/input.test.tsx
@@ -1,0 +1,7 @@
+import { Input } from '.';
+
+describe('Input', () => {
+  it('is truth', () => {
+    expect(Input).toBeTruthy();
+  });
+});