+++ src/components/layout/layout.test.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { render } from '@testing-library/react';
+
+import { Layout, LayoutProps } from '.';
+
+describe('Layout', () => {
+  it('is truthy', () => {
+    expect(Layout).toBeTruthy();
+  });
+});