+++ src/components/link/link.test.tsx
@@ -1,0 +1,7 @@
+import { Link } from '.';
+
+describe('Link', () => {
+  it('is truthy', () => {
+    expect(Link).toBeTruthy();
+  });
+});