+++ src/components/tag/tag.test.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { render } from '@testing-library/react';
+
+import { Tag, TagProps } from '.';
+
+describe('Tag', () => {
+  it('is truthy', () => {
+    expect(Tag).toBeTruthy();
+  });
+});