+++ src/components/text/text.test.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import { render } from '@testing-library/react';
+
+import { Text, TextProps } from '.';
+
+describe('Text', () => {
+  it('is truthy', () => {
+    expect(Text).toBeTruthy();
+  });
+});