+++ src/typings.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 /**
  * Default CSS definition for typescript,
  * will be overridden with file-specific definitions by rollup
@@ -7,11 +8,13 @@ declare module '*.css' {
   export default content;
 }
 
-interface SvgrComponent extends React.StatelessComponent<React.SVGAttributes<SVGElement>> {}
+// eslint-disable-next-line no-undef
+interface SvgrComponent
+  extends React.StatelessComponent<React.SVGAttributes<SVGElement>> {}
 
 declare module '*.svg' {
   const svgUrl: string;
   const svgComponent: SvgrComponent;
   export default svgUrl;
-  export { svgComponent as ReactComponent }
+  export { svgComponent as ReactComponent };
 }