+++ utils/tokens.ts
@@ -1,0 +1,39 @@
+export const tokens = {
+  colors: {
+    $buttonHover: `rgba(0, 0, 0, 0.05)`,
+    $buttonPressed: `rgba(0, 0, 0, 0.09)`,
+    $primaryButtonHover: `#0077FF`,
+    $primaryButtonPressed: `#0088FF`,
+    $primary: '#0099FF', // primary brand color
+    $secondary: '#e65f5c', // secondary brand color
+    $black100: '#000000', // Primary text color
+    $black200: '#444444', // Secondary text color
+    $black300: 'e0e0e0', // tertiary text color
+    $white: '#fff',
+    $border: '#dadce0', // border color for anything
+    $placeholder: '#708090', // placeholder color
+    $statusSuccess: `#28a745`,
+    $statusWarning: `#ff9a04`,
+    $statusCritical: `#e22121`,
+    $statusUnknown: `#757575`,
+  },
+  fontSizes: {
+    $h1: '2rem', // 32px
+    $h2: '1.75rem', // 28px
+    $h3: '1.5rem', // 24px
+    $h4: '1.25rem', // 20px
+    $h5: '1rem', // 16px
+    $h6: '0.75rem', // 12px
+  },
+  space: {
+    $1: `0.25rem`, // 4px
+    $2: `0.5rem`, // 8px
+    $3: `0.75rem`, // 12px
+    $4: `1rem`, // 16px
+    $5: `2rem`, // 32px
+    $6: `3rem`, // 48px
+  },
+  other: {
+    $boxShadow: `1px 1px 3px #dedede`,
+  },
+};